py/requirements.txt
flask==2.3.2
python-dotenv==1.1.0
flask-cors==4.0.0
gunicorn==21.2.0